Dont snapshot resolve after reject (#1964)
There is a simple issue here in that during an error, resolve() is
called after reject() is called. This is because 'close' is ALWAYS
called no matter what if the stream finishes, or there is an error.

You can verify this if you add console logs to the writer handlers in
the staging branch.