loads decryptedNotes without computing unconfirmed balance
unconfirmed balance is updated every time the decrypted notes store is updated,
so it isn't necessary to recompute unconfirmed balance.

calculating the balance on load has the benefit of ensuring that the balance
matches the decryped notes stroe when the node starts, but we read the balance
from the database each time we use it rather than tracking it in memory, so
we're already relying on updating the balance in tandem with the notes.