Try to actually parse a public address for validation (#2040)
Just checking that a public address is a hex string of 86 characters is not enough, because it might not be a valid subgroup point when broken down into the diversifier and transmission key pair. So by properly parsing it, we validate that it is indeed an actually valid public key.