Fix create account / scan race condition (#2079)
There is a bug where we don't wait for the create account transaction to
complete before we insert the account into the in memory cache, exposing
it to other code to access. This leads to situations where an account
is available to things like rescan, but haven't been added to the
database yet which we require. This leads to crashes where you can load
an account but it'll crash when the headHashes table for that account is
empty.

We need to wait for the TX to commit, before we expose it to in memory,
or set it as the default.