filters nullifiers and transactions by account on load
fixes transaction expiration with multiple accounts

expiring transactions on a wallet with multiple accounts may fail with `Error:
nullifierToNote mappings must have a corresponding decryptedNote`.

expiring transactions deletes the transaction from all accounts in the wallet.
each account loads all transactions and all nullifiers from the database, but
loads only the decrypted notes that it owns. the error occurs when a transaction
uses a nullifier to try to load a decrypted note that it does not own.

the solution suggested here is for each account to load only the nullifiers that
correspond to decrypted notes that the account owns. along with filtering
nullifiers by account it makes sense to filter transactions by account.

- adds test of expiring transactions on node with multiple accounts
- builds nullifierToNoteHash mapping while loading notes
- reduces the nullifiers tracked by an account to only those that map to a
  decrypted note that belongs to that account.
- filters transactions by account to only those transactions where an account
  either owns a note that was spent in the transaction or a note output by the
  transaction.
- adds tests of loading transactions and nullifiers by account