deserializes transaction on TransactionValue (#2113)
* deserializes transaction on TransactionValue

changes the `transaction` field of `TransactionValue` to an instance of the
`Transaction` primitive instead of a serialized transaction buffer. avoids
defining a separate anonymous type elsewhere in wallet and avoids using spread
operator to coax TransactionValue into that type.

* updates tests to use useMinersTxFixture

* Fix how TransactionValue encoding is tested

The issue is that it was using jest's equals, which JSON serializes
the object and then tries to check equality that way. That fails because
the non primitive Transaction type doesn't match when it's decoded
because the cached hash doesn't exist yet.

* Check in fixture

Co-authored-by: Jason Spafford <nullprogrammer@gmail.com>