Rewrite balance command and add --explain (#2243)
* Rewrite balance command and add --explain

I started this PR by trying to add an explain flag to balance. I now pass
through all the data used to calculate balance through the API so that it can be
explained. I'm half considering that these should be a different RPC, but we can
pass them through for now.

I started creating a new CurrencyUtils because the existing one is a mess.
Things that take amounts are taking numbers, which should be an immediate sign
that overflow errors are happening since amounts are bigint and numbers are
capped at 54 bits in Javascript. I can't go in and fix all the old places so I
figure we should just start creating a new one and delete the old ones in
another PR.

This also changes balance to not show the unconfirmed balance by default and
hides it behind the explain flag. Too many people were getting confused by
unconfirmed, which btw wasn't even unconfirmed it was your pending balance.

This PR now has 3 distinct balance definitions which are more correct:
 - confirmed: All notes on the main chain NOT in the confirmation range
 - unconfirmed: All notes on the main chain even if they are in the confirmation
   range
 - pending: All notes on the main chain or in transactions not mined yet

```
> ironfish accounts:balance
Account: jason (b500383)
Balance: $IRON 393.18339367
```

```
> ironfish accounts:balance --explain
Account: json (b500383)

Your balance is made of mined notes that are safe to spend
Balance: $IRON 393.18339367

0 notes worth $IRON 0.00000000 are mined but are within 12 blocks of head of the chain
Unconfirmed: $IRON 393.18339367

0 notes worth $IRON 0.00000000 are waiting on miners to mine them
Pending: $IRON 393.18339367
```

* Fixed tests