Move account:transaction to its own command (#2256)
* Move account:transaction to its own command

Moved most of the logic out of accounts:transactions into its own
command. I feel like that command was overloaded and not doing what
you expect anyway. Iexpect when I pass a TX hash for it to display the
same output, but filtered on only that transaction. Instead it has a
totally alternate flow.

* Fixed name and description