Fix staging test failures (#2260)
* Convert SocketAdapter to listen using events

This is the pattern we use in connecting and other clients and
adapters.

* Properly set started to false

* Dont reject server on error

The documentation says this error is passed in the case that it was not
listening to begin with, which is not actually an error.

* Use destroy() not end() to close client

end() does not close a connection, it sends a finish packet to the
server cleanly severing the connection. The problem is its still writing
data in this case, and when the connection is closed soon after it'll
error because were still reading / writing data to the socket.

* Null out the client after close

* Change test timeout to 20s

AFAIK binding to port 0 is very expensive on github.