Dont materialize unspent notes when creating a tx (#2299)
* Reduce default min confs to 5

* Clarify message on why TX cannot be created

I got this error and checked that I was synced with the block chain. The
error is misleading because it says your node must be synced, but your
node can be synced and still get this error because it has nothing to do
with being synced.

* Dont materialize unspent notes when creating a tx

This loaded all TX when creating a note and joinin its related
transaction to get its sequence.