compacts chain db before creating snapshot (#2360)
service:snapshot is failing sporadically with ENOENT errors. what seems to be
happening is that between the start of the 'tar.create' and the end of it, files
are removed from the database. LevelDB compaction may be running in the
background and merging recent chain database changes.

manually execute database compaction prior to zipping the database to try to
prevent compaction during the zip process.