allows users to set datadir using env variable
updates datadir flag to read from 'IRONFISH_DATA_DIR' env variable. allows users
to set this env variable to avoid having to pass '--datadir=...' with every
command if they use a non-default datadir.

datadir is somewhat unique as a flag: it can't be set in an Iron Fish
config file, since the config file is in the datadir.