Transaction uses 1 spender key instead of taking in the same key multiple times (#2395)
* Add spender SaplingKey to Transaction, use for .receive()

This commit introduces a change to take the spender's SaplingKey as part of the ProposedTransaction constructor, and uses that key when `.receive()` is called. This removes the need to pass in the key for the receive function. The following commits will use it for .post and .spend as well.

* Remove spender SaplingKey argument from Transaction `post()`

This makes the post function simpler by using the (now included) Transaction's spender SaplingKey

* Remove spender SaplingKey argument from Transaction `spend()`

* Fix up napi binding test