stop wallet rebroadcasting expired transactions (#2399)
we rebroadcast any transactions in the wallet that have not been added to a
block. however, now that we continue to store expired transactions in the wallet
we must ensure that we do not rebroadcast them.

rebroadcasting expired transactions also updates them in the wallet db with the
current head as submittedSequence. when this update is made they are added to
the wallet db store for pending transaction hashes. they then 're-expire' and
their notes are reverted to unspent (even if they have later been spent in
another transaction).