Refactor SpendParams, OutputParams to be builder-style
The goal of this PR was to push the heavy logic into the `.post`/`.build` functions, so a transaction can be crafted without a lot of computation, such as generating the proofs. This also required some changes to the `ProposedTransaction` as well, since it expected the proofs to be built when they are added. By moving all of the logic into `.post`, we were able to de-duplicate some computation, as well as split up the proof-generation, signatures and spend-signing into cleaner steps within the `.post` function.