refactor(rust): SpendParams, OutputParams to be builder-style (#2417)
* Refactor SpendParams, OutputParams to be builder-style

The goal of this PR was to push the heavy logic into the `.post`/`.build` functions, so a transaction can be crafted without a lot of computation, such as generating the proofs. This also required some changes to the `ProposedTransaction` as well, since it expected the proofs to be built when they are added. By moving all of the logic into `.post`, we were able to de-duplicate some computation, as well as split up the proof-generation, signatures and spend-signing into cleaner steps within the `.post` function.

* missed some renames during rebase

* Re-add verify_proof sanity check

* Add some tests for the miners fee code paths in MerklePath and OutputBuilder

* Convert an .unwrap into a properly handled Result

While not strictly necessary, it's probably best to avoid unwraps and expects unless absolutely necessary. If this code gets refactored and introduces a bug here, we could be introducing panic behavior instead of just handling the error more gracefully.