reuses transaction size for different fee rates
estimates fees for all fee rate priority level with only one call to
'createSpends' to avoid iterating over notes multiple times.

changes 'getPendingTransactionSize' to return both the size and the spender
change for a pending transaction without accounting for the fee.

implements 'getPendingTransactionFee' to use the size and spender change to
determine the fee based on how big the transaction is and whether more notes or
spends are needed to cover the fee and spender change.

implements 'estimateFees' to return estimates for all fee rate priority levels
and updates estimateFee RPC endpoint.

accounts for additional note for spender change in transaction sizes.