adds accounts:repair command
adds a cli command intended to fix balance and note indexing issues in the
wallet without the need for a full rescan/reset of the wallet.

repairs the balance by recomputing it from 0 and adding all unspent notes.
uses the chain to determine if each note has been spent and changes to spent if
the note is marked as unspent

repairs sequenceToNoteHashes by ensuring that all notes that are on chain have
the sequence of the transaction that created the note. any sequence to note hash
mapping that does not have a note is removed.

repairs nonChainNoteHashes by removing any note hashes for notes that are on
chain.

repairs nullifierToNoteHash by removing any nullifier that maps to a note that
is not in the wallet or maps to a note that is not on the chain.