fixes updating of head hashes
call chainProcessor remove directly

update wallet before disconnecting blocks because we need to be able to load the
blocks from the chain in order to disconnect them from the chainProcessor

call updateHeadHash directly for each account since updateHeadHashes will not
update any account that is not up to date. since we're going backwards, the
accounts are not up to date after the first removal.

use progress bar for wallet updates