avoids re-syncing expired transactions in wallet
if we receive a transaction message from a peer for a transaction that has
already expired we will sync that transaction and add its notes to the wallet.
since the transaction record in the database has not changed we do not add the
transaction to pendintTransactionHashes and it never re-expires in the wallet.
this results in the balance updating to include notes that do not exist.

updates syncTransaction to check whether an off-chain transaction is expired
before syncing it.

if the transaction has a blockHash but would other be expired, such as when
syncing during a re-org, we still sync the transaction.

adds tests of both cases.