uses config tempDir for backup files
the backup command writes files to a temporary directory before uploading to S3.
instead of generating a new directory on each command invocation we can use the
tempDir specified by the config, which is located within the datadir.

the config tempDir makes backup consistent with other ironfish commands and
makes the write location for backup consistent across invocations.