fixes backup command multipart uploads (#2633)
* fixes backup command multipart uploads

the backup command uses an S3 multipart upload to copy a zip of the chain and/or
accounts databases to an S3 bucket.

this command fails if the AWS credentials aren't set properly in the S3Client.
if there are no credentials set, we need not pass a 'credentials' object to the
constructor.

makes accessKeyId and secretAccessKey flags interdependent. makes both flags
automatically read from env variables if not passed in.

fixes parsing of credentials to pass no credentials to the S3Client constructor
if no credentials are passed to the command or set in the environment.

this won't allow uploads without access keys, however, as S3Client requires some
credentials. if none are passed in it will read them from AWS credentials
configured on the machine.

* uses aws cognito to allow unauthenticated uploads

if aws credentials are not set or passed in aws aws cognito to create an
        identity and fetch credentials for it.

created a cognito identity pool in aws and use the id of that pool here for
creating identities.

adds @aws-sdk/client-cognito-identity to dependencies

* adds default bucket to backup command

* Revert "adds default bucket to backup command"

This reverts commit 2d1249ceb5285c7f1fbdda8161fc650c5bc838c6.