Fix KeyStore not saving value if matching default
There is a bug here where if we load all the keys, we add them to a set
to keep track of those keys. That set is used to determine which keys
get saved when we call keyStore.save(). It's also used by
keyStore.isSet(). Since we don't add the key to the set when we call
keyStore.put(), if it matches the default value it's dropped, and
isSet() will always return false.