adds separate method for syncing pending transactions
pending transactions are currently added to the account via the same function,
syncTransaction, that we use to process transaction when blocks are added to and
removed from the chain. this can make transaction processing confusing.

creates a separate 'addPendingTransaction' method for transactions created by an
account or gossiped to the node. it only adds a transaction to an account if the
account has not processed the transaction before.

the addTransaction method on the account
- creates new decryptedNote records for each output note in the transaction
  received by the account
- adds each of these notes to nonChainNoteHashes
- marks each spent note as spent
- adds the transaction to pendingTransactionHashes
- creates new record in walletDb transactions store
- updates balance once with the delta from received notes and spent notes

these changes also update tests and fixtures to use addPendingTransaction
instead of syncTransaction.