feat(ci): Cache rust build on staging (#2732)
Due to the limitations of cache access (https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#restrictions-for-accessing-a-cache), every PR following a change to rust code will have to rebuild until that change reaches master. If we build on every push to staging and master, we can make sure the rust cache is updated for staging as well.