fixes wallet handling of fork transactions from other wallets
when a block is disconnected from the chain, the transactions on that block are
pending: they may be included in a future block connected to the chain.

the wallet marks notes as spent when they are spent in pending transactions.
when a block is disconnected, those notes are still marked as spent _unless_ the
transaction was created by another wallet.

this is incorrect because the transaction is still pending even if a different
wallet created it. marking the spent notes as unspent could result in the wallet
creating double-spend transactions.

these changes fix this issue by only marking spent notes as unspent when a
transaction expires.

- changes processTransactionSpends to only mark spends as spent, never unspent
- updates test to enforce correct handling of forked transactions from other
  wallets