wallet should not scan when all accounts up to date
if all accounts in the wallet are up to date (account head hash equal to chain
processor head hash), then there are no blocks to scan. we implement a property
on the wallet, shouldRescan, that checks this. however, scanTransactions does
not check this when initializing a scan. if a scan starts with beginHash equal
to endHash, then that block will be scanned again because
chain.iterateBlockHeaders uses an inclusive range. this leads to adding the
block to the account a second time.

- checks shouldRescan within scanTransactions before beginning a scan
- adds unit test
- fixes genesis test that masked this issue