adds chain state fields to balances datastore
instead of storing only a bigint with the asset balance as the value in the
balances store, stores an object containing the balance and the block hash and
sequence of the most recent balance update.

now that we only update the balance when blocks are connected and disconnected
from the chain the stored balance value corresponds to a specific chain state.
storing this chain state alongside the balance will allow us to inform users of
what chain state their balance reflects. a number of users were confused by
balances while accounts were scanning during phase 2: their transactions would
be on the chain, but the transaction wouldn't be reflected in their account
balance because the account was behind the chain.

- defines BalanceValue type to include balance, blockHash, and sequence
- replaces value in balances datastore with BalanceValue
- propagates BalanceValue through all reads/writes on balances store