removes balance repair from accounts:repair
now that an account's stored balance only reflects transactions on chain it is
not feasible to recompute the balance only from an account's decrypted notes. we
should not include spent notes in the balance, but we don't currently have an
efficient way to determine if the note was spent in a transaction that is on the
chain or in a transaction that is pending.

we can still make some repairs to the spent attribute stored on the note, so the
old implementation of `repairBalance` processes notes without updating balance
and is renamed `repairNotes`.

future changes to better support recomputing the balance:
- add sequence and block hash to balance store to allow incremental updates to
  balance
- add balance delta to transactions to avoid iterating over all notes and spends
- add an index from sequence to transaction hash to support efficiently
  iterating over transactions in order
- add an index from nullifier to transaction hash to support identifying whether
  a note was spent in an on-chain transaction