adds balance delta mapping to transactionValue
users currently cannot see how much they sent or received in each transaction
listed in the CLI. the only way to present this information to the user with the
current walletDb schema would be to load each note and spend for each
transaction each time we look up the transaction.

stores the balance delta for each asset on the value in the wallet transactions
store.

serializes buffer map as a list of key-value pairs prepended with the length of
the list. keys are 32-byte asset identifiers and values are bigints.