adds account filtering logic to connect/disconnect
the wallet connectBlock method should only update accounts where the block to
connect follows the account head. similarly, the disconnectBlock method should
only update accounts whose head is equal to the block to disconnect.

instead of filtering accounts outside of these methods and passing a list of
accounts in, the methods should execute the filtering themselves to ensure that
they don't leave account data in a bad state.

- only connects a block for an account if the account's head hash is equal to the
  block's previousHeadHash or if the block is the genesis block and the account
  has a null head hash
- only disconnects a block for an account if its hash is equal to the account's
  head hash
- simplifies scan logic: do not track accounts during the scan, but rely on
  connectBlock to only connect blocks where appropriate