changes headHashes to store both hash and sequence
the walletDb headHashes store tracks the hash of the latest block that has been
synced to each account. we often use that hash to look up the sequence for that
block in order to identify the earliest or latest head hash in the wallet.
storing the sequence in the same datastore eliminates a chain dependency from
the wallet.

- renames headHashes to headers
- defines HeaderValue and NullableHeaderValueEncoding; headers keeps the same
  behavior as headHashes of storing null if the account has not scanned any
  blocks