makes note decryption for spender optional
the decryptNotes worker task first attempts to decrypt a note for the owner
using an account's incoming view key and then tries to decrypt it for the
spender using the account's outgoing view key.

the wallet does not store notes decrypted for the spender. these notes are
skipped completely when updating the walletDb.

by making decryptForSpender optional and skipping this decryption step when
syncing transactions we can expect a ~2x speedup in scanning transactions in the
wallet.