simplifies getTransactions logic with limit
before we indexed transactions by time we implemented a feature on the
getTransactions endpoint to sort transactions and limit the number returned.

the implementation iterated over all transactions and maintained a priority
queue of the 'n' most recent transactions.

now that we iterate over transactions in order of time it's much simpler to
impose the limit by stopping early when we've streamed enough transactions back.

deletes the 'handle*' functions added to distinguish between the 'limit' and
'all' cases and uses the limit to return early from streaming all transactions.