adds IRON amount to wallet:transactions output (#2907)
* adds IRON amount to wallet:transactions output

the transactionValue stored in the walletDb now includes a mapping from assetId
to the balance delta for that account and asset in the transaction. this allows
us to return transaction amounts in our RPC responses and display them in the
CLI.

serializes assetBalanceDeltas as a JSON string because yup does not support
map types.

only includes the balance delta for IRON in the CLI output table. we can show
balance deltas for all assets in a transaction more easily in the output for a
single transaction in wallet:transaction

* changes serialization of tx assetBalanceDeltas

uses Array<{ assetId: string; delta: string }> instead of a JSON string of a map
object

* Changed loop to a find

Co-authored-by: Jason Spafford <nullprogrammer@gmail.com>