shows all assets for each transaction in transactions list
adds additional rows for each asset in a transaction beneath the native asset
row.

each custom asset row displays the asset ID and the net amount of that asset in
the transaction.

converts '*Count' columns to 'extended' columns that are not displayed by
default.

updates getTransactions and getTransaction to include asset name in
assetBalanceDeltas with each transaction so that asset name can be displayed
with balances.

truncates the asset name column unless the '--extended' flag is passed to show
all columns in the table.