fixes calculation of confirmed balance
the confirmed balance for an account is equal to the unconfirmed balance of the
account excluding any unconfirmed transactions.

we currently subtract the value of any notes received in unconfirmed
transactions, but we do not add the value of any notes spent. this leads to a
temporarily incorrect value computed for confirmed balance.

- fixes calculateUnconfirmedCountandConfirmedBalance to use transactions and
  transaction balance deltas
- changes CLI terms to refer to unconfirmed transactions instead of unconfirmed
  notes
- adds test confirming bug and fix