fixes account reset to use mark and sweep (#3050)
the purpose of resetting an account is to clear all of an account's data but
preserve the account's keys and the record of the account.

this is functionally equivalent to deleting all account data except the data in
the accounts store. however, this will crash for large accounts because we
attempt to delete all data within a single database transaction.

instead of deleting the account outright we can use a mark and sweep strategy:
'resetAccount' marks an account as deleted so that its data is cleaned up
asynchronously in the event loop.

to prevent deletion of the account record 'resetAccount' creates a new account
with the same name and same key but a different account id. the id is not user
facing and is only used for database storage.

- defines resetAccount
- splits the logic of removeAccount into removeAccount and removeAccountByName
- deletes the implementation of Account.reset