adds transactionAmounts datastore to walletDb
the transactions datastore includes a field 'assetBalanceDeltas' on each record
that stores a mapping from asset ID to the net amount or balance delta for that
asset in the transaction.

however, there are some cases, such as computing the amount of an asset that an
account has available to spend, where it is useful to separate the amount that
the account input into the transaction from the amount that the account received
output from the transaction.

the transactionAmounts datastore stores the input and output amounts for each
account, transaction hash, and asset id.

this datastore will allow us to efficiently compute the 'available' or
'spendable' balance by reading the amount spent in a transaction without
iterating over a transaction's spends and looking up decrypted notes for those
spends in the walletDb.