feat(ironfish): introduce an event loop for mining pool
The goal here is to simplify the pool logic a little bit. With the
coming refactors, it will be much easier if we have the main logic in
one place, so it can be extended and refactored without introducing a
bunch of random timeouts or intervals. This also will make it easier to
read and maintain the code going forward, I hope.