fix: pool nextPayoutAttempt timestamp is taken before the function runs
This allows the function to mimic existing functionality, where the
payouts are generally at a fixed time, regardless of the time taken for
the payout to execute. It is subtly different in that only 1 will
execute at a time. If one takes longer than the interval, it will simply
execute the next one immediately after it ends. This is a temporary
change, since this logic is going to be removed once payouts can be made
on-demand and deterministically.