feat(ironfish): mining pool tracks status of blocks its submitted
This introduces a new table to the mining pool database, block. The pool
will insert a row every time it successfully submits a block. The event
loop will check any unconfirmed blocks, and update their status
accordingly. For now, that status includes whether it's on the main
chain, and whether it's confirmed. This gives us the ability to track
how much reward was earned within a specific timeframe.