fixes syncing transactions with no change (#3211)
a recent change (#2897) made decrypting spender notes optional. a side effect of
that change is that the wallet no longer syncs transactions to an account if the
account does not receive any notes from the transaction.

this is correct unless the account spent notes in the transaction and did not
receive any notes in the transaction.

fixes syncing transactions without change received by checking whether an
account owns any of the spends in the transaction.

syncs a transaction to an account if the account has decrypted output notes for
the transaction OR the account owns any of the notes that were spent as inputs
to the transaction.