sets connected transaction timestamp to block header timestamp
the wallet currently sets the timestamp field of a record in the transactions
datastore to the time that the wallet first synced that transaction.

however, if the wallet rescans transactions, then these timestamps will all be
set to a time during the rescan, and the relative time of transaction is lost.

instead we can set the timestamp of the transaction equal to the timestamp in
the block header when that transaction is connected to the chain.

reported in #3219