feat(ironfish): add new payout share table
This table is functionally identical to the existing one, but gives us a
foreign key to payoutPeriod table, which will eventually be used to
create new payout transactions. When that is done, this table will have
an additional foreign key relation to a payout transaction table

The old table/logic will be removed in one complete removal PR, rather
than piece-meal removal, so that we don't accidentally get anything into
a broken state or miss something.