fixes little-endian encodings in walletDb
we store sequences in several walletDb datastores: sequenceToNoteHash,
sequenceToTransactionHash, and pendingTransactionHashes

we use little-endian encodings in the keys for each of these stores. the problem
is that we rely on iterating over the keys of these stores *in order* so that we
can load keys using sequence ranges. we cannot use little-endian encodings for
this kind of ordering; we need to use big-endian encoding instead.

- adds UInt32EncodingBE for big-endian encoding of Uint32 values
- replaces datastores using little-endian encodings with new ones using
  big-endian encodings
- adds data migration to replace the old stores with the new stores
- adds tests of WalletDB methods that rely on ordered iteration