feat(ironfish): new payout logic for mining pool
This introduces the core logic needed to create payouts with the new
additions to the codebase. Specifically, it now knows which shares have
been paid out, how much reward was earned in a payout period, whether
its blocks and transactions are confirmed and whether transactions need
to be undone in the database if they expire.

It lacks some of the polish of the existing payout logic, which will be
moved over and tweaked as needed as we begin to remove the old logic.