adds unspentNoteHashes store to walletDb
the unspentNoteHashes datastore is an index on note hashes that supports
efficiently looking up the values of notes that are confirmed in a given
sequence range and have not been spent in another transaction.

this query pattern supports computing the 'available' balance for an account and
asset by summing the values of all notes that are confirmed and unspent.

the datastore uses a five-part compound key:
- account prefix
- asset ID
- sequence
- value
- note hash

the stored value is always null

the datastore is updated according to these rules:
- when a pending transaction is created

        - remove entries for any note that is spent in the transaction

- when a transaction is connected

        - add entries for all output notes (if they have not been spent in a
          pending transaction)
        - remove entries for any note that is spent in the transaction

- when a transaction is disconnected

        - remove entries for all output notes

- when a transaction expires

        - add entries for any notes spent in the expired transaction

this store will be used to calculate the 'available' or 'spendable' balance that
an account has for a given asset. to calculate the balance:
- determine the maximum sequence for confirmed notes
- iterate over all keys in unspentNoteHashes in the range:

        - gte: account prefix, asset id, sequence = 1
        - lt: account prefix, asset id, sequence = max sequence + 1

- sum the value stored in the fourth part of each key