feat(ironfish): add new share concept to mining pool (#3316)
* feat(ironfish): add new share concept to mining pool

This is functionally very similar to the existing shares. The only
differences are that it is associated to a payout period and a payout
transaction. This way, we know which payout period it belongs to, and
which transaction it belongs to. This also allows us to make sure that a
share is paid out again in the event of a payout transaction expiring.

The old logic/table will be removed in a follow-up commit that removes
all of the old stuff in one change, to minimize disruption to the
existing pool and make it easier to transition to the new logic.

* fix: Use an `IN` clause in the markSharesPaid function on publicAddress