Fix ASCII encoding errors
There are many places where we write convert UTF8 to ascii that could be
causing problems. The default encoding to bufio is 'binary', which is an
alias to 'latin-1' which is effectively ASCII. This will change our
types to force users to specify the encoding to avoid issues like this.
I also changed our encodings over to use UTF-8.