Mining pool refactors (#3353)
* test(ironfish): introduce some test scaffolding for the mining pool (#3265)

* test(ironfish): introduce some test scaffolding for the mining pool

This introduces some basic scaffolding for the poolDatabase and
MiningPoolShares classes. While this change doesn't introduce any meaningful unit
tests, it will enable testing with the refactor PRs, so that we can be
more confident in the changes being made.

* fix: poolShares test with an in-memory sqlite db

* feat(ironfish): Add explicit payout period to mining pool (#3268)

* feat(ironfish): Add explicit payout period to mining pool

This adds a "payout period" with a fixed duration, such that any blocks or
shares found during that time will be tied to that period. This will
allow the pool to calculate the exact amount to payout each share in a
deterministic fashion. Materializing explicit rows instead of relying on
timestamps allows the code to be a bit more explicit and easier to
reason about.

* fix: add database transaction when rolling over payout periods

* feat(ironfish): mining pool tracks status of its submitted blocks (#3275)

This introduces a new table to the mining pool database, block. The pool
will insert a row every time it successfully submits a block. The event
loop will check any unconfirmed blocks, and update their status
accordingly. For now, that status includes whether it's on the main
chain, and whether it's confirmed. This gives us the ability to track
how much reward was earned within a specific timeframe.

* feat(ironfish): mining pool tracks status of transactions its created (#3285)

This introduces a new table which allows the pool to insert a row
whenever it makes a new transaction. The event loop will update the
status of any unconfirmed and unexpired transactions. This allows the
pool to know what transactions it has made, as well as giving it the
ability to re-pay out transactions in the event of one expiring.

* feat(ironfish): add new share concept to mining pool (#3316)

* feat(ironfish): add new share concept to mining pool

This is functionally very similar to the existing shares. The only
differences are that it is associated to a payout period and a payout
transaction. This way, we know which payout period it belongs to, and
which transaction it belongs to. This also allows us to make sure that a
share is paid out again in the event of a payout transaction expiring.

The old logic/table will be removed in a follow-up commit that removes
all of the old stuff in one change, to minimize disruption to the
existing pool and make it easier to transition to the new logic.

* fix: Use an `IN` clause in the markSharesPaid function on publicAddress

* feat(ironfish): new payout logic for mining pool (#3322)

* feat(ironfish): new payout logic for mining pool

This introduces the core logic needed to create payouts with the new
additions to the codebase. Specifically, it now knows which shares have
been paid out, how much reward was earned in a payout period, whether
its blocks and transactions are confirmed and whether transactions need
to be undone in the database if they expire.

It lacks some of the polish of the existing payout logic, which will be
moved over and tweaked as needed as we begin to remove the old logic.

* use pool's config for transaction expiration delta

* minor improvements

* refactor(ironfish): Mining pool notifications updated with the new logic (#3336)

* refactor(ironfish): Remove old pool logic and test logic we kept (#3337)

* refactor(ironfish): Remove old pool logic and test logic we kept

* remove old tables and fix names of some migrations

* feat(ironfish): Allow the pool migrations to be silenced during tests (#3350)

* fix(ironfish): Mining pool deletes shares that cannot be paid out (#3358)