refactor(ironfish): move assembleBlockFromResponse fn
This moves `assembleBlockFromResponse` function from the `peerNetwork`
into the `blockFetcher`. It is a helper function that has no dependence
on the `peerNetwork` class, and is only used by the `blockFetcher`
class, we might as well have it live where it is used. The `peerNetwork`
class is already pretty bloated, so this is a small step to alleviate
that.