feat(ironfish): improve the limits of GetBlockHeaders (#3431)
* feat(ironfish): improve the limits of GetBlockHeaders

This introduces a limit of 1024 headers that can be requested per
GetBlockHeaders request. A block header is currently ~180 bytes, which
would make the max message size around 185kb. This also introduces a
limit to the number of total block headers that can be looked up from
disk, given that GetBlockHeaders can use a `skip` amount to potentially
use an excessive amount of disk usage.

* fix: off-by-1 error when counting lookups and include tests

* move header lookup into a const in versions