returns undefined for missing nullifierToNoteHash
when loading values from nullifierToNoteHash we convert the return value from
undefined to null if the key is not found in the datastore. we do not do this
for any other datastore in the walletDb.

returns undefined when key is not found in the datastore to be consistent with
other stores in the walletDb.