handles insufficient funds for one or more fee rates
we use Promise.all to parallelize createTransaction requests for all three fee
rates (slow, average, fast) when neither a fee nor fee rate are specified in the
send command.

however, if the sender account does not have enough funds to pay the fast fee
rate, then the whole transaction fails. the user cannot select the slow fee rate
even if their account has enough funds to pay that fee.

handles insufficient balance errors when creating transactions for different fee
rates. inserts an 'Insufficient funds' option into the inquirer prompt and
throws the error if that option is selected.

removes parallelization of the three fee rate requests to simplify error
handling. Promise.allSettled is an alternative that would preserve
parallelization, but makes it more difficult to distinguish the type of error
when a request fails.